style: format oscillator header

+++ include/orpheus/dsp/oscillator.hpp
@@ -42,10 +42,12 @@ constexpr double sine_taylor(double angle) noexcept {
   return x - (x3 / 6.0) + (x5 / 120.0) - (x7 / 5040.0) + (x9 / 362880.0);
 }
 
-template <std::size_t Size> constexpr std::array<double, Size> make_sine_table() noexcept {
+template <std::size_t Size>
+constexpr std::array<double, Size> make_sine_table() noexcept {
   std::array<double, Size> table{};
   for (std::size_t i = 0; i < Size; ++i) {
-    const double phase = (static_cast<double>(i) / Size) * 2.0 * std::numbers::pi;
+    const double phase =
+        (static_cast<double>(i) / static_cast<double>(Size)) * 2.0 * std::numbers::pi;
     table[i] = sine_taylor(phase);
   }
   return table;
@@ -289,9 +291,9 @@ private:
     PinkState pink{};
   };
 
-  [[nodiscard]] double render_voice(VoiceState& voice, Waveform waveform, double phase_increment,
-                                    double pulse_width, double sub_increment,
-                                    double& sub_mix) noexcept;
+  [[nodiscard]] double render_voice(VoiceState& voice, Waveform waveform,
+                                    double phase_increment, double pulse_width,
+                                    double sub_increment, double& sub_mix) noexcept;
 
   void apply_phase_sync_if_needed() noexcept;
 
@@ -303,7 +305,8 @@ private:
                                             std::size_t voice_index) noexcept;
   static void advance_phase(double& phase, double increment) noexcept;
 
-  template <typename T> static constexpr T lerp(T a, T b, T alpha) noexcept {
+  template <typename T>
+  static constexpr T lerp(T a, T b, T alpha) noexcept {
     return a + (b - a) * alpha;
   }
